add v1 folder for versioning and index, store and show method for tasks
