+++ laravel-app/app/Models/Task.php
@@ -8,4 +8,6 @@ use Illuminate\Database\Eloquent\Model;
 class Task extends Model
 {
     use HasFactory;
+
+    protected $fillable = ['title','scheduled_for','parent_id'];
 }
